+++ lib/compressors/clean-css.js
@@ -10,9 +10,9 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var Promise = require('bluebird');
 var CleanCSS = require('clean-css');
+var utils = require('../utils');
 
 /**
  * Expose `compressCleanCSS()`.
@@ -30,7 +30,7 @@ module.exports = compressCleanCSS;
 function compressCleanCSS(settings, data) {
   return new Promise(function(resolve) {
     var result = new CleanCSS(settings.options).minify(data).styles;
-    fs.writeFileSync(settings.fileOut, result, 'utf8');
+    utils.writeFile(settings.fileOut, result);
     resolve(result);
   });
 }
