+++ lib/compressors/csso.js
@@ -10,9 +10,9 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var Promise = require('bluebird');
 var csso = require('csso');
+var utils = require('../utils');
 
 /**
  * Expose `compressCSSO()`.
@@ -30,7 +30,7 @@ module.exports = compressCSSO;
 function compressCSSO(settings, data) {
   return new Promise(function(resolve) {
     var result = csso.minify(data, settings.options.restructureOff);
-    fs.writeFileSync(settings.fileOut, result, 'utf8');
+    utils.writeFile(settings.fileOut, result);
     resolve(result);
   });
 }
