+++ lib/compressors/gcc.js
@@ -10,8 +10,8 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var path = require('path');
+var utils = require('../utils');
 var execCompressor = require('../runner');
 
 /**
@@ -33,10 +33,9 @@ module.exports = compressGCC;
  */
 
 function compressGCC(settings, data) {
-  return execCompressor(gccCommand(settings), data, settings)
+  return execCompressor(gccCommand(settings.options), data, settings)
     .then(function(data) {
-      fs.writeFileSync(settings.fileOut, data, 'utf8');
-      return data;
+      return utils.writeFile(settings.fileOut, data);
     })
     .catch(handleErrors);
 }
@@ -52,21 +51,7 @@ function handleErrors(err) {
  * Google Closure Compiler command line.
  */
 
-function gccCommand(settings) {
-  var args = [];
-  var options = settings.options;
-
-  // TODO : refactoring
-  Object.keys(options).forEach(function(key) {
-    if (options[key] && options[key] !== false) {
-      args.push('--' + key);
-    }
-
-    if (options[key] && options[key] !== true) {
-      args.push(options[key]);
-    }
-  });
-
+function gccCommand(options) {
   return [
     '-server',
     '-XX:+TieredCompilation',
@@ -75,5 +60,5 @@ function gccCommand(settings) {
     binGcc,
     '--warning_level',
     'QUIET'
-  ].concat(args);
+  ].concat(utils.buildArgs(options));
 }
