+++ lib/compressors/no-compress.js
@@ -10,8 +10,8 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var Promise = require('bluebird');
+var utils = require('../utils');
 
 /**
  * Expose `noCompress()`.
@@ -28,7 +28,7 @@ module.exports = noCompress;
 
 function noCompress(settings, data) {
   return new Promise(function(resolve) {
-    fs.writeFileSync(settings.fileOut, data);
+    utils.writeFile(settings.fileOut, data);
     resolve(data);
   });
 }
