+++ lib/compressors/sqwish.js
@@ -10,9 +10,9 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var Promise = require('bluebird');
 var sqwish = require('sqwish');
+var utils = require('../utils');
 
 /**
  * Expose `compressSqwish()`.
@@ -30,7 +30,7 @@ module.exports = compressSqwish;
 function compressSqwish(settings, data) {
   return new Promise(function(resolve) {
     var result = sqwish.minify(data, settings.options.strict);
-    fs.writeFileSync(settings.fileOut, result, 'utf8');
+    utils.writeFile(settings.fileOut, result);
     resolve(result);
   });
 }
