+++ lib/compressors/uglifyjs.js
@@ -10,9 +10,9 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var Promise = require('bluebird');
 var uglifyJS = require('uglify-js');
+var utils = require('../utils');
 
 /**
  * Expose `compressUglifyJS()`.
@@ -32,7 +32,7 @@ function compressUglifyJS(settings, data) {
     var options = settings.options;
     options.fromString = true;
     var result = uglifyJS.minify(data, options);
-    fs.writeFileSync(settings.fileOut, result.code, 'utf8');
+    utils.writeFile(settings.fileOut, result);
     resolve(result.code);
   });
 }
