+++ lib/compressors/yui.js
@@ -10,8 +10,8 @@
  * Module dependencies.
  */
 
-var fs = require('fs');
 var path = require('path');
+var utils = require('../utils');
 var execCompressor = require('../runner');
 
 /**
@@ -36,8 +36,7 @@ module.exports = compressYUI;
 function compressYUI(type, settings, data) {
   return execCompressor(yuiCommand(type, settings.options), data, settings)
     .then(function(data) {
-      fs.writeFileSync(settings.fileOut, data, 'utf8');
-      return data;
+      return utils.writeFile(settings.fileOut, data);
     });
 }
 
@@ -46,24 +45,11 @@ function compressYUI(type, settings, data) {
  */
 
 function yuiCommand(type, options) {
-  var args = [];
-
-  // TODO : refactoring
-  Object.keys(options).forEach(function(key) {
-    if (options[key] && options[key] !== false) {
-      args.push('--' + key);
-    }
-
-    if (options[key] && options[key] !== true) {
-      args.push(options[key]);
-    }
-  });
-
   return [
     '-jar',
     '-Xss2048k',
     binYui,
     '--type',
     type
-  ].concat(args);
+  ].concat(utils.buildArgs(options));
 }
