+++ lib/setup.js
@@ -14,6 +14,7 @@ var fs = require('fs');
 var path = require('path');
 var glob = require('glob');
 var extend = require('xtend');
+var utils = require('./utils');
 
 /**
  * Default settings.
@@ -42,7 +43,7 @@ module.exports = setup;
 function setup(inputSettings) {
   checkMandatories(inputSettings);
 
-  var settings = extend(clone(defaultSettings), inputSettings);
+  var settings = extend(utils.clone(defaultSettings), inputSettings);
   settings = extend(settings, wildcards(settings.fileIn, settings.publicFolder));
   settings = extend(settings, setPublicFolder(settings.fileIn, settings.publicFolder));
 
@@ -125,14 +126,4 @@ function mandatory(setting, settings) {
   if (!settings[setting]) {
     throw new Error(setting + ' is mandatory.');
   }
-}
-
-/**
- * Clone an object.
- *
- * @param {Object} obj
- */
-
-function clone(obj) {
-  return JSON.parse(JSON.stringify(obj));
 }
